Implemented GSTrans'd variables; added simple tests.
The GSTrans of the type 0 variables is now performed in the SetVarVal
SWI. The set and read are now tested by a small BASIC program which
uses the module SWIs (not the OS SWIs). There is a new *Command which
prints the variables out.

The buffer is global so there will be reentrancy issues - deal with
this later.